# Improve the performance of parsing text