--- a/benchmark/html/index.js
+++ b/benchmark/html/index.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = {
+  parsing: require("./parsing")
+};