--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -2,7 +2,8 @@
 
 module.exports = {
   jsdom: require("./jsdom"),
-  dom: require("./dom")
+  dom: require("./dom"),
+  html: require("./html")
 };
 
 require("./prepare-suites")("", module.exports);