--- a/lib/jsdom/living/nodes/CharacterData-impl.js
+++ b/lib/jsdom/living/nodes/CharacterData-impl.js
@@ -33,10 +33,10 @@ class CharacterDataImpl extends NodeImpl {
     }
 
     if (offset + count > length) {
-      return this._data.substring(offset);
+      return this._data.slice(offset);
     }
 
-    return this._data.substring(offset, offset + count);
+    return this._data.slice(offset, offset + count);
   }
 
   appendData(data) {
@@ -62,8 +62,8 @@ class CharacterDataImpl extends NodeImpl {
       count = length - offset;
     }
 
-    const start = this._data.substring(0, offset);
-    const end = this._data.substring(offset + count);
+    const start = this._data.slice(0, offset);
+    const end = this._data.slice(offset + count);
 
     this._data = start + data + end;