@Zirro 2018-11-24
I added a benchmark which can be used to tell if we regress or improve further on this in the future.