@domenic 2018-12-02
Commented at https://bugs.chromium.org/p/v8/issues/detail?id=6730#c4 and will tweet for some extra attention. https://bugs.chromium.org/p/v8/issues/list?can=2&q=slice+substring&colspec=ID+Type+Status+Priority+Owner+Summary+HW+OS+Component+Stars&x=priority&y=owner&cells=ids has some other vaguely-interesting issues.