@mathiasbynens 2018-12-06
The fix for this just landed in V8: https://chromium-review.googlesource.com/c/v8/v8/+/1365282